Move api routes to their own file
