+++ src/services/apiRoutes.ts
@@ -1,0 +1,10 @@
+const ROUTES = {
+  GET_NOTIFICATIONS: '/MatchesForReview',
+  GET_PATIENT_DOCUMENT: '/Document',
+  GET_GOLDEN_ID_DOCUMENTS: '/GoldenRecordDocuments',
+  UPDATE_NOTIFICATION: '/NotificationRequest',
+  CREATE_GOLDEN_RECORD: '/Unlink',
+  LINK_RECORD: '/Link'
+}
+
+export default ROUTES
